pyproject.toml
[tool.poetry]
name = "djano_project_template"
version = "0.1.0"
description = "Django project template"
authors = ["Dmytro Parfeniuk"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "==3.1.4"
django-environ = "==0.4.5"
django-cors-headers = "==3.5.0"
django-anymail = "==8.2"
django-storages = "==1.11.1"
django-celery-beat = "==2.0.0"
djangorestframework = "==3.12.2"
djangorestframework-simplejwt = "*"
drf-yasg = "==1.20.0"
Pillow = "==8.0.0"
pytz = "==2020.4"
ipdb = "==0.13.4"
psycopg2-binary = "==2.8.6"
celery = "==4.4.7"
redis = "==3.5.3"
uvicorn = "==0.12.1"
gunicorn = "==20.0.4"
boto3 = "==1.16.56"
djoser = "==2.1.0"

[tool.poetry.dev-dependencies]
flake8-isort = "==4.0.0"
black = "==20.8b1"
pytest-django = "==4.0.0"
pytest-mock = "==3.5.1"
pytest-cov = "==2.11.0"
mixer = "==7.1.2"
ptvsd = "==4.3.2"
rope = "==0.18.0"
django-extensions = "==3.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
